driver: support on-demand sensors
We want to support two different types of sensors:

- Periodic (period_ns != 0): Produces data with a regular period.
- On-demand (period_ns == 0): Produces data on-demand, when the user requests a
  read.

Previously, there was no way to tell the driver when we want to produce data, so
on-demand sensor data didn't really work. So we add a next() call to the driver
interface, which tells the driver to produce a datum for a given data ID. next()
will be automatically called from the synchronous hound_read() interface for any
on-demand data requested by the context. It will *not* be called by
hound_read_async(), as next() is blocking.