driver: make driver API private
Make the driver API private so that we can change it freely without breaking
anyone. In order to still allow for driver registration, each driver exports a
driver-specific registration function. This has the side-benefit of not exposing
the user to void * nastiness and making the register call have only the
paremeters that are relevant for the given driver.

The implication of all this is that we don't allow for third-party drivers; all
drivers must be in the source tree or risk ABI-breakage without warning. This is
a good default, as it allows future flexibility. We could always change to allow
third-party drivers, but moving the other way is much harder.