TODO: remove item
Unfortunately, there's not really an easy way to stop the drivers from touching
the seqno field in a hound_record. That said, it's completely harmless if they
do so, as we call parse before we set seqno.

The only two methods I can think of to do this are:
- Add an internal struct inside the public hound_record definition. This allows
  the drivers to see only the internal struct but forces the public definition
  to be messier on account of it. This seems like a bad tradeoff.

- Create another struct that encompasses only the fields we want drivers to see
  (all but seqno). Then, cast to that struct before you pass a record into the
  driver. Although this would work, it raises the possibility of bugs when the
  definitions change, or when we change compilers or compiler options and the
  struct gets packed differently. We can static_assert to guard against some
  cases (like changing definitions), but not all (like changing memory layouts).

Oh well.