ctx: fix ref_driver_map error handling
When ref_driver_map fails to ref a driver, it attempts to unref all
drivers proceeding this driver in the map. However, it does so by
assuming it can do iter-- and iter++ on a hash map, which is not true,
since the entries are contiguous. D'oh!

Fix this by properly iterating the hash map.