io: fix accidentally huge static section
We noticed plugin loading failing on the raspberry pi due to running out
of memory. Turns out the .bss section for libhound is huge. This is
caused by the accidentally huge POLL_BUF_SIZE of 1GB, which should have
been 1MB but was off by a factor of 1024. Given that poll returns pretty
quickly, let's set this to 100KB instead.