driver: refactor datadesc contract
Currently, we have no way to support driver data descriptors that are
dynamically allocated. To fix this, tweak the contract so that drivers
allocate the driver descriptors directly, and the driver core owns the
memory and frees the descriptors when the driver is unregistered. This
has the drawback that drivers with static descriptors still have to
allocate memory. In exchange, the driver interface is simpler because
drivers don't need to implement a free_datadesc call. I think this is
the right tradeoff because most non-trivial drivers (real drivers rather
than test drivers) need to allocate descriptors because they probe the
system to generate the descriptors.

Since driver descriptors are now dynamically allocated, additionally
change them to be in contiguous memory, to reduce fragmentation.