driver: add per-instance context
So far, we have gotten away with keeping drivers as singletons. This is
OK for simple drivers, but IIO will need multiple instances of the same
driver, and similar situations will probably keep happening. So, we want
the ability for a driver to set and later get back custom void *
context. Although we could do this by passing the context into each
function, that clutters up the prototypes and incurs a tax even on
drivers that don't need it.

So, add a context pointer to the driver struct, and make a thread-local
pointer to cache the active driver struct. This pointer gets set prior
to calling into a driver operation, so the driver can rely on the
correct driver struct being present during operation. When the driver
calls drv_ctx(), it gets back the context specific to itself rather than
another instance of the same driver. The driver can call drv_get_ctx()
to set the context pointer for its own driver struct.