io: properly set driver context
The I/O layer was not setting the driver context prior to calling the
driver parse function, so the context was NULL, breaking drivers. We
didn't see this bug because, although the nop driver tests contexts, it
doesn't test them from the parse function.

Fix this by splitting out the driver ops code and calling it from the
I/O layer. Add a test case for this in the nop driver too.