driver: remove queue only if fd is still present
Currently, when derefing a driver, we first remove the driver queue and
then the driver fd. There is a race window between removing the queue
and the fd in which data is generated by the I/O layer but has nowhere
to go, so it leaks.

Fix this by removing the driver queue only if we didn't remove the fd
from the I/O layer entirely. If we removed the fd, we destroy all the
queues associated with that fd anyway, so we don't need to -- and must
not -- remove the queue from that fd.