driver/iio: restart properly in the error case
If we call setdata and the device is active, we need to restart the
device. Currently, we do so only in the success case. This means that if
we stop the device and then hit an error, we fail to restart the device.

Fix this by restarting the device when needed in both the success and
error cases.