iio: fix handling of negative values
Currently, the copy function doesn't handle negative values properly, as
it uses the combination of a union and int_fast{8,16,32,64}_t types.
Since the fast types don't have to be the same width as the regular
types, the union cast can go awry.

Fix this by using regular casts and no fast types.