gps: assert that int is 4 bytes
gpsd uses a raw int in its struct gps_fix_t mode field, but int is not
guaranteed to be 4 bytes. Since we want to guarantee an ABI to the user,
we fix it at 4 bytes. Thus if the assumption of a 4-byte int ever
changes, we will need code to handle it.