add schema parser
In order to make Hound handle self-describing data, we need to make an
API to describe the binary blobs you get from a Hound record. Although
we can hardcode the description into each driver, it is more robust and
easier to manage if we instead encode it into a schema. That way,
calling programs will know what to expect and, since the schema is
version-controlled, we can control changes to it.

To handle the schema, we have two choices:
(a) Code-generate C code to back the schema (ugly but fast at runtime).
(b) Parse the schema when we load each driver.

I prefer option (b), as code generation is brittle and hard to maintain.
And, since the schemas are not long, the runtime cost of parsing them is
negligible. That said, we really don't want runtime failures. So, we add
a meta-schema that describes the format of the schema itself. Then, we
can check the schema at compile-time and trust the schema at runtime.