refactor data IDs to be 32-bit
We want to accomodate drivers that are not fixed-function (in other
words, where the data IDs the drivers can generate are not known at
compile-time). Example of these are bus protocols like OBD-II, in which
the driver can request many different types of data. In order to support
this, let's use a 32-bit unsigned int instead of an enum to represent
data IDs. We could use a UUID, but since all hound drivers are checked
into the codebase, we can make sure everything is unique without
expanding the data IDs to 128 bits.