schema: new format
Use a new format for the schemas with several improvements:

- Use a list instead of a map. This way, we don't have to specify or
  check offsets.

- Specify the schema name in the schema instead of in the code. This
  seems obvious but that's not the way it was being done (probably a
  holdover from before we had schemas).

- Don't specify the length in the schema, as the type tells you all you
  need to know, so we just risk redundancy.

Note that although we don't have offset or length in the schema, they
get auto-calculated by the driver core, so the public API is unchanged.
Although offset and length are technically redundant, they could save
time for API users, so we keep them.

Since we're changing the parsing code substantially, we might as well
also take this opportunity to convert to libyaml document-based parsing,
which is much easier to read and maintain than state-based parsing.
Goodbye confusing state machine!