io: lock driver ops
Surprisingly, we were not taking the driver lock during parse
operations. Most of the time, this lock should be a no-op, as it's
uncommon that the consumer is trying to call stop, reset, next, etc. at
the same time. But, since we guarantee that drivers don't have to write
reentrant code, we should honor the contract.

To do this, we break the driver lock into two locks:

- The ops lock, used only to lock driver_ops_* operationss and guarantee
  that drivers don't have to be reentrant.
- The state lock, used for the driver core manipulating the rest of the
  driver struct.

Note that if the driver core changes something that affects the I/O
thread (e.g. adding fds or queues, or changing drv->fd), we still need
to pause the I/O thread to prevent conflicts.