io: fix poll race using ppoll
In unlucky cases, we are hitting the race condition that caused
ppoll/pselect to exist.  Specifically, if PAUSE_SIGNAL is sent after
io_wait_for_ready() returns but before poll() is entered, the signal is
handled and poll() will block forever, since the signal is not sent
again.

The solution is to block PAUSE_SIGNAL and use ppoll, which unblocks the
signal for the duration of the call. If PAUSE_SIGNAL was sent prior to
starting ppoll(), it will get processed, since the signal got queued up
while blocked.

Note it's very rare to hit this issue, and I only saw it when adding
unrelated debug prints.