io: push only data matching ID
Currently, there's a bug: if you subscribe to one data ID that a driver
generates, you get everything that the driver generates. I probably
missed this because in the past, we mostly had only drivers generating
one type of data. IIO changed this and brought the bug to light.

Fix this by letting the I/O layer know what data ID a queue is
associated with. In this way, we can deliver only data explicitly
requested.