io: support pull mode scheduling
Currently, all drivers are implicitly push mode: after being configured,
they arrange for data to be pushed, but they don't have to request data.
Pull-mode drivers instead have to periodically request data instead of
having it pushed to them. An example pull-mode driver is interacting
with a request-response bus, or a client-server model.

Add a notion of push vs. pull-mode scheduling to the I/O and driver
layer. A given driver picks its scheduling mode. For push mode,
everything works as it correctly does. For pull mode, the I/O layer
calls driver next() at the configured the frequency. Thus the user
shouldn't have to care if a driver is push vs. pull, as either way, it
will honor the configured frequency.

In order to exercise this new functionality, we make the test counter
driver pull-mode.