ctx: stop calling ctx_next from ctx_read
For contexts containing more than 1 data request, calling ctx_next
automatically from ctx_read is problematic. The issue can be seen from
the example of a context with 2 data requests calling hound_read(ctx,
1). This will end up calling ctx_next(..., 1), which will generate *2*
records, not *1*. Thus *by design*, asking to read a single record will
generate more than 1 record, and doing this in a loop will cause the
receive fd to eventually overflow. In fact, for on-demand data, ctx_read
will always call ctx_next, and this is unavoidable.

Now, we could do something smarter and not allow reading a single
record... instead, force callers to read a multiple of the number of
on-demand data points. However, this risks issues when a context has a
mix of on-demand and periodic data, as it's completely valid to read a
single bit of data. This becomes quite messy.

Fix this issue by just decoupling ctx_next from ctx_read. They now need
to be called separately, and calling hound_read on an entirely on-demand
context will block forever. ctx_next must instead be called first.