can: fix period setting
Because the period is being set *after* allocating the context, the
second time we run this function, the period is still 0. The test is
still passing because of code inside ctx_read calling ctx_next behind
the scenes, but it's *not* actually setting the driver to periodic mode.

Fix this stupid bug.