io: fix memory leak
In commit a7daa9 ("io: push only data matching ID"), we accidentally
created a memory leak because we no longer push each record into all the
queues (instead pushing only to queues matching ID. However, we
initialize the refcount to be the number of queues. This shortcut worked
when we put a record into every queue, but now that we put it into only
some queues, we need to do an explicit incref each time we put into a
queue.