make CAN driver an OBD-II driver
We have a CAN driver, which just processes raw CAN frames. However, CAN
is a bus, not a data type, so this always felt a little bit hacky.
Astute readers may have noticed that the CAN unit tests sent OBD-II CAN
frames, as this was the use-case I had in mind when writing the CAN
driver. I wanted to write an OBD-II driver, but I wasn't sure how to
represent the data for it in hound's data IDs, as OBD-II is an open
universe of (mode, pid) combinations instead of fixed, hardcoded IDs.

Now that we've refactored the data IDs to allow open universes of IDs,
we can now support OBD-II. We represent the PIDs by packing them into a
32-bit hound data ID like this:

0xffMMPPPP

ff is required padding.
MM is 8 bytes for the mode.
PPPP is 16 bytes for the PID.

Using this mechanism, we can do 1:1 translation for hound data ID and
(mode, pid) combo.

Note further that OBD-II processing is handled by the yobd library,
which is now a dependency of hound. yobd has its own schema system, so
there's now a script to convert from a yobd schema (describing valid
modes and pids) to a hound schema (describing hound data IDs for the
same modes and pids). As a convenience, I've checked in an invocation of
this script into schema/deploy, but whenever the yobd schema changes, we
should re-run the script and check it in again.

As the number of drivers/libraries grows, I'll eventually make it
possible to make each driver a compile-time option.

In order to test the OBD-II driver, create an OBD-II simulator, which
responds to valid OBD-II requests with randomized data. Run this
simulator in a separate thread inside the unit tests.