schema: fix memory leak
Currently, we're not properly freeing all schema descriptors.
Specifically, when we free the schema descriptors that aren't mentioned
in the driver's datadesc call, we aren't freeing the name strings. For
example, if an iio driver advertises an accelerometer but no gyroscope,
then the gyroscope schema descriptor wouldn't get freed.

Fix this, and refactor the destruction to occur in one place and
eliminate code duplication.