schema: allow for fixed length bytes type
Currently, "bytes" always means a variable length format. We should
allow for fixed length formats (e.g. 3 bytes of arbitrary data), as well
as for variable length, as long as the format is the last in the format
list. Variable length formats must be last in the list so that they're
parseable and it's always clear where one format ends and the next
begins.

Add a "size" field to the schema, which is required when the type is
"bytes" and otherwise not present. Then, make the driver code set "size"
only when the type is something other than "bytes" (in other words, a
primitive type in which the size is known).