io: use self-pipe trick instead of signal
Currently, we use signals to pause the poll loop. This is a fragile
because anything on the system can use our magic SIGUSR1 signal and
pause the poll loop. We're also stealing the signal from the process
using this library, which in to optimal.

Instead use the self-pipe trick, where we make a pipe to ourself and
write to it when we want to pause the poll loop. Notably, this means the
first fd in the poll loop is the special self-pipe, and the index math
for fds needs to be modified to fix this. Now fds have an index one
greater than fds. In order to protect ourselves from bugs relating to
this, encapsulate all the index math into functions.

For more info on the self-pipe trick:
https://lwn.net/Articles/177897/