io: make sequence numbers truly per-queue
Currently, sequence numbers are attached to an fdctx. This has a bug: If
a context subscribes to multiple drivers, the sequence numbers won't be
monotonically increasing.

Fix this by making each queue track its own sequence numbers. By keeping
track of the sequence number assigned to the front and back items in the
queue, the queue can update sequence numbers when items are pushed,
popped, or drained.

Unfortunately, this requires changing the callback interface because
sequence numbers cannot be put into records or record_info's anymore.
This is because sequence numbers are now different for each queue, but
the records are shared among queues. Thus the sequence number must be
given as a separate parameter in the user callback, which ultimately
comes from the queue pop operation.