ctx: allow blocking reads to be interrupted
Currently, there's no way to stop a blocking read. This causes an issue
if an application wants to stop a context (say, to shutdown) but
hound_read has not returned, perhaps because a driver is not creating
new data.

Address this by making hound_read able to return HOUND_CTX_STOPPED if a
context is stopped while a blocking read is active. Note this means we
need to add another parameter to hound_read to indicate to the caller
how many records were actually read, as the reading will stop when
HOUND_CTX_STOPPED is returned.