drivers: refactor schema contract
Currently, drivers have to provide a schema file and then allocate
descriptors that match that schema. If they fail to do so, the driver
core will error out. This is kind of silly, as the drivers are
duplicating some of what the schema parsing has already done.

Restructure the schema contract such that:
- A driver-declared schema file gets parsed
- Drivers are given the schema descriptors from this file
- Drivers dynamically determine which descriptors are available on this
  system, and at what frequencies they are available.

This reduces code duplication, makes it easier to write a driver, and
eliminates error cases.