mqtt: handle new mosquitto init/cleanup semantics
In mosquitto versions 1.6.10 and higher, init/cleanup became refcounted
and can be safely called multiple times. Prior to this version, we need
to assume that the application initializes mosquitto prior to
initializing this MQTT driver, as if we call init/cleanup ourselves, we
will cause memory leaks and/or corruption.

Add MQTT driver code to check the version and do the right thing, and
update the driver documentation accordingly.