io: unify pull and push-mode drivers
Add the ability for a driver to specify a timeout in its poll call,
which the MQTT driver needs in order to call mosquitto_loop_misc once a
second or so. This also lets us unify push and pull-mode drivers such
that the basic driver primitive is push, and we can get rid of the
concept of scheduling modes. Basically, drivers can write their own poll
function, or they can use the default provided push-only or pull-only
functions.