io: pause polling during ctx modify
When we are modifying a context, we previously had a race condition in
which data could be lost after we removed an fd/queue associated with
data and then added it back. Previously, we solved this by draining the
fd before shutting it down. However, this trick works only with
pull-mode drivers, as push-mode drivers could trigger an infinite loop.
Plus, it slows down the modify operation and is aesthetically "gross".

This means we need a new solution for the brave new world of mixed-mode
drivers unified by a poll() operation. The solution is that, during a
modify operation, we need to:

- Pause polling during the entire operation
- Don't do a driver stop/start, as that would remove the fd and lose
  data.

With this solution, we don't lose data, no matter the driver type!