io: close possible memory leak
Close a possible (though unlikely) memory leak, in case a driver uses
drv_push_records outside of a poll operation, *and* a user context
associated with that driver is being modified concurrently. In this
case, a driver may produce a record info that has no backing queue, and
we need to free that record info as we have no place to put it.