io: protect io_push_records with a rwlock
In case io_push_records gets called by a driver outside of a poll()
function, we need to protect this with a lock, as we need to grab an fd
ctx, so we can't have the fdctx array changing at the same time. We can
use a read-write lock, as the poll loop and io_push_records just read
from the array, so they can have safe concurrent access. The
io_{add,remove}_{fd,queue} functions modify arrays, so they need
exclusive access.