io: take read lock in need_to_pause
Since we are touching s_ios.fds, we should take the read lock here. Note
that technically, the chance of having issues here is extremely low, as
no one ever modifies the first index of s_ios.fds (the self-pipe).
However, if someone appends to s_ios.fds, xvec could perform a realloc,
causing the memory behind s_ios.fds to be rewritten or change address
and leading to an invalid memory read.