io: set last_pull to before the pull, not after
We're approximating the last pull time for the fd as a whole. We're
looping over an array of data IDs, so setting the last pull time to
either the first or the last pull is not fully correct, so we may as
well approximate using the first, as we have to grab that timestamp
anyway. In this way, we can call get_time_ns() just once instead of
twice.