ctx: refactor driver maps
The driver maps make an implicit assumption that each driver has either
only periodic data types or only on-demand data types. However, this
assumption isn't really valid anymore, so we should refactor the driver
maps.

Refactor the maps so that we have:
- Driver data map: drv --> list of active requests
- On demand map: drv --> data ID

The driver data map is the one used for driver ref/unref, and the
on-demand map is just an optimization to make ctx_next faster.