driver: simplify the modify operation
Currently, modify is rather confusing because it uses ref/unref, while
modify shouldn't really have to do a ref/unref. This causes confusing
code flow (having a modify flag to the ref/unref operators that
fundamentally changes what the functions do) and it's also inefficient
because it means we have to stop the polling loop, add/remove queues,
etc.

Refactor modify so that we don't do a ref/unref, and we pause polling
only if we are modifying an active driver and need to change the queues
associated with it.