driver/mqtt: stop disabling the data callback
Previously, we had to disable the MQTT data callback while waiting for a
subscribe/unsubscribe reply because, during a modify operation, there
would be no queue associated with this driver. With the latest driver
modify refactoring, this is no longer the case, and there is always a
queue associated with an active driver. Thus, we can keep the MQTT data
callback enabled and never drop data!