driver: fix error case in driver_ref
Currently, in driver_ref, we don't restore the active driver data if we
fail later on after setting the driver data. Fix this, and also fix a
related bug in which we call drv_op_setdata on the *new* data rather
than *all* the active data. In our tests, these are one and the same,
but that's not guaranteed in general.